+++ src/EntityGraphQL/EntityGraphQLArgumentException.cs
@@ -2,7 +2,7 @@
 
 namespace EntityGraphQL
 {
-    public class EntityGraphQLArgumentException : Exception, IExposableException
+    public class EntityGraphQLArgumentException : Exception
     {
         public EntityGraphQLArgumentException(string message) : base(message) { }
         public EntityGraphQLArgumentException(string parameterName, string message) : base($"{message} (Parameter '{parameterName}')") { }