+++ src/tests/EntityGraphQL.Tests/ErrorTests.cs
@@ -10,7 +10,7 @@ namespace EntityGraphQL.Tests
         public void MutationReportsError()
         {
             var schemaProvider = SchemaBuilder.FromObject<TestDataContext>();
-            schemaProvider.AddMutationsFrom<PeopleMutations>(new SchemaBuilderMethodOptions() {  AutoCreateInputTypes = true });
+            schemaProvider.AddMutationsFrom<PeopleMutations>(new SchemaBuilderMethodOptions() { AutoCreateInputTypes = true });
             // Add a argument field with a require parameter
             var gql = new QueryRequest
             {
@@ -308,6 +308,24 @@ namespace EntityGraphQL.Tests
             Assert.NotNull(results.Errors);
             Assert.Single(results.Errors);
             Assert.Equal("Field 'people' - Error occurred", results.Errors[0].Message);
+        }
+
+        [Fact]
+        public void QueryReportsError_AllowedExceptionAttribute()
+        {
+            var schemaProvider = SchemaBuilder.FromObject<TestDataContext>(new SchemaBuilderSchemaOptions { IsDevelopment = false });
+            // Add a argument field with a require parameter
+            var gql = new QueryRequest
+            {
+                Query = @"{
+                    people { error_Allowed }
+                }",
+            };
+
+            var testSchema = new TestDataContext().FillWithTestData();
+            var results = schemaProvider.ExecuteRequest(gql, testSchema, null, null);
+            Assert.NotNull(results.Errors);
+            Assert.Equal("Field 'people' - This error is allowed", results.Errors[0].Message);
         }
     }
 }