+++ src/tests/EntityGraphQL.Tests/SortTests/SortTests.cs
@@ -168,7 +168,7 @@ namespace EntityGraphQL.Tests
             Assert.Equal("Zoo", person.lastName);
             var schemaType = schema.Type("PeopleSortInput");
             var fields = schemaType.GetFields().ToList();
-            Assert.Equal(12, fields.Count);
+            Assert.Equal(13, fields.Count);
             Assert.Contains("people(sort: [PeopleSortInput!] = [{ height: ASC }]): [Person!]", schema.ToGraphQLSchemaString());
         }
         [Fact]